refactor: Update fetching of dashboard json and refactor function distrubtion
